[ig] kops-generated ASG name as a module output
This way, you'll be able to interact with ASGs created by kops from the
rest of your Terraform code. That convenient - for instance - to attach
an ELB to an autoscaling group on which some containers will listen on
host ports.

In particular, doing so for the ingress controller nodes is a good way
to make sure requests get routed directly to a node that runs the
ingress controller, without having to go through kube-proxy.