Add possibity to create k8s PVs on top of EFS
For those requiring ReadManyWriteMany persistent volumes on their AWS
Kubernetes clusters (to scale old, dirty & maxi-stateful PHP apps, for
instance), this commit adds the ability to create persistent volumes on
top of an EFS volume. Note that you'll need to create the EFS volume
yourself (with Terraform if you want to).