Enhance the test suite a little bit
We're getting a step closer to a highly available setup (both on the
cloud provider level and on the kubernetes layer).

What this imply is that we'll also be able to test that karch templates
and creates clusters that are resilient to node failure, remain
available during rolling upgrades, etc.