Allow fully private cluster topologies
This commit add the possibility to create fully private cluster
topologies, where even the internal Route53 zone is private as well as
the kubernetes master ELB.

Of course, one still wants to access his/her cluster :) For that, a
peering connection with a VPC containing a VPN appliance (such as
OpenVPN, Pritunl or Pertino) must be created by the end user.

Our/kops provisioning process require access to the cluster during the
bootstrap phase. Therefore, the ID of the route tables (and subnets) in
the Kubernetes cluster must be accessible before we can declare our
cluster as "up". For that reason we need to create these resources via
Terraform before running the "kops create cluster" command, hence the
many additions in the "aws/cluster/networking.tf" file.

For the record, we're creating as many /21 CIDR-ed public subnets as
availabilty zones spanned by the cluster (2048 allocatables IPs per
public subnet) and support up to 8 AZs. We're creating the same number
of /19 private subnets with 8192 allocatable IPs each. This should
cover most use cases :)

NOTE: the bastion ELB is still created in a public subnet. Making it
optional may be possible in the future, however, leaving an emergency
entrypoint (CIDR filtered and protected by key authentication) seemed
like a wise thing.