[breaking][aws] Make it possible to use IAM roles
Currently, the "provider" configuration in Terraform is left to the
end user. However, kops use the local AWS configuration, unless
overidden by an "aws-profile-env-override" variable.

To improve developer experience, this variable has been replaced with
a simple "aws-profile" env. variable that can be used to select any
profile in `~/.aws/credentials` or `~/.aws/config`.

Note that providing an IAM role (e.g. on a self-hosted CI/CD pod) may
require to set an empty string for this variable.