Make a default security group settable for our LBs
For LBs created as part of a LoadBalancer kubernetes services, this
commit adds the ability to set a security group for this LB.

What this makes possible is to apply this SG to both nodes and LBs
and allow "self" traffic on the k8s node port range (30000-32767).
This way, load balancers (in the same VPC) can target services without
the need to add a the whole VPC CIDR range to the node's security
group.