[BREAKING] Update for Terraform v0.12
THIS IS A MAJOR BREAKING UPDATE, USING THIS NEW VERSION OF THE KARCH
MODULE REQUIRES TO UPGRADE TERRAFORM TO VERSION V0.12.0 AT LEAST.

This commit makes the karch module compatible with Terraform v0.12.0.
------

It also adds the ability to associate VPCs with the created Route53 zone
in which the Kubernetes cluster lives. That's useful, for instance, to
bind this zone with a VPN VPC.

This is necessary since the AWS provider now imposes that all VPC
associations are created within the "aws_route53_zone" block.

Bofore, people were using the external aws_route53_zone_association
resource, which is strongly not recommended any more:
https://www.terraform.io/docs/providers/aws/r/route53_zone_association.html

NOTE: THIS COMMIT BREAKS BACKWARD COMPATIBILITY WITH TERRAFORM
VERSIONS PRIOR TO V0.12.0